# lib/api-zod/src/generated/types/generateIdeasResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateIdeasResponseIdeasItem } from "./generateIdeasResponseIdeasItem";

export interface GenerateIdeasResponse {
  ideas: GenerateIdeasResponseIdeasItem[];
}