# lib/api-zod/src/generated/types/generateScriptBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateScriptBodyTensionLevel } from "./generateScriptBodyTensionLevel";
import type { GenerateScriptBodyVideoLength } from "./generateScriptBodyVideoLength";

export interface GenerateScriptBody {
  topic: string;
  channelStyle?: string;
  videoLength: GenerateScriptBodyVideoLength;
  tensionLevel: GenerateScriptBodyTensionLevel;
  tensionTechniques?: string[];
  inspirationLinks?: string;
  model?: string;
}